general improvements
changed tickers.txt to tickers_USDT.txt
added XRP to tickers
fixed error when closing the bot
changed signal names

+++ src/helpers/scripts/sell_all_coins.py
@@ -33,5 +33,7 @@ def sell_all():
                     f"{timestamp} Sell: {coins[coin]['volume']} {coin} - {buy_price} - {last_price} Profit: {profit:.2f} {price_change:.2f}%")
 
     # delete files who store current coins bought
-    os.remove('../test_mode_coins_bought.json')
-    os.remove('../coins_bought.json')
+    if os.path.exists('../test_mode_coins_bought.json'):
+        os.remove('../test_mode_coins_bought.json')
+    if os.path.exists('../coins_bought.json'):
+        os.remove('../coins_bought.json')
